Fix query parameters handling on API requests
This introduces a mechanism for resolving the query parameters stored in the
args dictionary in all actions. These optional parameters were previously not
passed along with the API request in the K8sClient and were not resolved.
The Jinja template for the action script files has been modified to include
a params dictionary within the args holding the query parameters key-value
pairs. The k8s base client now consumes the params dictionary and passes it
as data for the URL query string upon making the request to the API.