Add query params and non-JSON responses handling to actions/client
This adds the latest changes to all actions and to the k8s client which have
to do with the introduced mechanisms for handling query parameters on making
requests to the k8s API and resolving correctly the API responses which are not
JSON-compliant but are, for instance, in plain text. The actions have been
re-generated from the modified templates and replaced accordingly.